% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INEI.R
\docType{data}
\name{PBI2007}
\alias{PBI2007}
\title{Base de datos de PBI a precios constantes 2007 por Departamento.}
\format{Base de datos en formato XTS con la data de Sistema de Informacion Regional para la Toma de Decisiones SIRTOD}
\usage{
data(PBI2007)
}
\description{
Base de datos de PBI a precios constantes 2007  del Sistema de Informacion Regional para la Toma de Decisiones SIRTOD [http://iinei.inei.gob.pe/iinei/SIRTOD/]
}
\keyword{datasets}
