% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INEI.R
\docType{data}
\name{PEA}
\alias{PEA}
\title{Base de datos de PEA (Poblacion Economicamente Activa) por Departamento.}
\format{Base de datos en formato XTS con la data de Sistema de Informacion Regional para la Toma de Decisiones SIRTOD}
\usage{
data(PEA)
}
\description{
Base de datos de PEA (Poblacion Economicamente Activa) por Departamento  del Sistema de Informacion Regional para la Toma de Decisiones SIRTOD [http://iinei.inei.gob.pe/iinei/SIRTOD/]
}
\keyword{datasets}
